Crate pmsi_rust

pmsi_rust: du PMSi avec Rust

pmsi_rust est une bibliothèque pour le PMSi en Rust basée sur Polars. Elle permet de lire, découper et requêter les fichiers PMSi ainsi que les exporter vers des fichiers de formats plus ouverts.

Table des matières

Premier exemple

Lire les RSA

Lire les RUM

Lire les autres fichiers

Premier exemple

let file_to_read = "~/Documents/data/mco/290000017.2022.12.ano";
let annee = "2022";
let extension = "ano";

use pmsi_rust::core::parse_pmsi;
use pmsi_rust::core::utils::io::write_dataframe_to_somefile;

let ano = parse_pmsi(file_to_read.to_owned(), 
                     annee.to_owned(), 
                     extension.to_owned()).unwrap();
 
let out_parquet: String = "~/Desktop/290000017.2022.12.ano.parquet".to_string();
write_dataframe_to_somefile(ano.clone(), out_parquet);

let out_csv: String = "~/Desktop/290000017.2022.12.ano.csv".to_string();
write_dataframe_to_somefile(ano.clone(), out_csv);

let out_json: String = "~/Desktop/290000017.2022.12.ano.json".to_string();
write_dataframe_to_somefile(ano, out_json);

Écriture de ~/Desktop/290000017.2022.12.ano.parquet
Écriture de ~/Desktop/290000017.2022.12.ano.csv
Écriture de ~/Desktop/290000017.2022.12.ano.json

Lire les RSA

Partie fixe uniquement

let file_to_read = "~/Documents/data/mco/290000017.2022.12.rsa";
let annee = "2022";
let extension = "rsa";
 
use pmsi_rust::core::parse_pmsi;

let rsa = parse_pmsi(file_to_read.to_owned(), 
                     annee.to_owned(), 
                     extension.to_owned()).unwrap();

Partie fixe et parties variables

let file_to_read = "~/Documents/data/mco/290000017.2022.12.rsa";
let annee = "2022";
 
use pmsi_rust::core::mco::parse_rsa;

let rsa = parse_rsa(file_to_read.to_owned(), 
                     annee.to_owned()).unwrap();

Lire les RUM

Partie fixe uniquement

let file_to_read = "~/Documents/data/mco/290000017.2022.12.rss.ini.txt";
let annee = "2022";
let extension = "rss.ini.txt";
 
use pmsi_rust::core::parse_pmsi;

let rum = parse_pmsi(file_to_read.to_owned(), 
                     annee.to_owned(), 
                     extension.to_owned()).unwrap();

Partie fixe et parties variables

let file_to_read = "~/Documents/data/mco/290000017.2022.12.rss.ini.txt";
let annee = "2022";
 
use pmsi_rust::core::mco::parse_rum;

let rum = parse_rum(file_to_read.to_owned(), 
                     annee.to_owned()).unwrap();

Lire les autres fichiers

Il y a une seule fonction pour tous les autres fichiers avec seulement une partie fixe.

let file_to_read = "~/Documents/data/mco/290000017.2022.12.dmi.txt";
let annee = "2022";
let extension = "dmi.txt";
 
use pmsi_rust::core::parse_pmsi;

let dmi = parse_pmsi(file_to_read.to_owned(), 
                     annee.to_owned(), 
                     extension.to_owned()).unwrap();

Il faut indiquer l’extension, parmi :

rsa

ano

ano.txt

rss.ini.txt

med

med.txt

porg

dmip

dmi.txt

iium

rsa_um

rsa_actes

rum_actes

pie

diap

Modules

cli

pmsi_rust en ligne de commande

core

Exemple de lecture de fichiers RSS et écriture en .parquet